agrego los sectores junto con las realizaciones de un espectaculo y sigo avanzando en el envio de alertas

+++ src/main/java/com/grupo6/persistence/repository/UsuarioRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.grupo6.persistence.model.Usuario;
 
 @Repository
-public interface UsuarioRepository extends BaseRepository <Usuario, Integer>{
+public interface UsuarioRepository extends BaseRepository <Usuario, Long>{
 
 	Optional<Usuario> findByEmail(String email);
